% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test.R
\name{bayes.wilcox.test}
\alias{bayes.wilcox.test}
\title{Bayesian First Aid Alternative to the t-test}
\usage{
bayes.wilcox.test(x, y, cred.mass = 0.95, mu = 0, n.iter = 30000,
  alternative = NULL, conf.level)
}
\arguments{
\item{x}{numeric vector of data values}

\item{y}{numeric vector of data values to be compared to x}

\item{cred.mass}{he amount of probability mass that will be contained in
reported credible intervals. This argument fills a similar role as
conf.level in \code{\link{wilcox.test}}}

\item{mu}{a number specifying an optional parameter used to form the null
hypothesis.}

\item{n.iter}{The number of iterations to run the MCMC sampling.}

\item{alternative}{ignored, only retained in order to maintain compability
with \code{\link{wilcox.test}}}

\item{conf.level}{identical to cred.mass,
ignored, only retained in order to maintain compability with
\code{\link{wilcox.test}}}
}
\value{
A list of class \code{bayes_wilcox_test}. It can be further inspected
 using the functions \code{summary}, \code{plot}, \code{diagnostics}
 and \code{model.code}.
}
\description{
Function implements a bayesian alternative to the wilcoxon.test() function.
Written as an addition to the Bayesian First Aid (BFA) package by
Rasmus Bååth (https://github.com/rasmusab). BFA's structure is largely
followed. S3 functions to be implemented as in BFA:
print, plot, summary, model.code, and diagnostics.
Method follows http://andrewgelman.com/2015/07/13/dont-do-the-wilcoxon/, and
fits a normal model using JAGS after rank-transformation.
}
