% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test.R
\name{diagnostics.bayes_paired_wilcox_test}
\alias{diagnostics.bayes_paired_wilcox_test}
\title{Plots and prints diagnostics regarding the convergence of the model.}
\usage{
diagnostics.bayes_paired_wilcox_test(x)
}
\arguments{
\item{x}{The output from a paired sample Bayesian Wilcox test model.}
}
\description{
Plots and prints diagnostics regarding the convergence of the model.
}
