% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test.R
\name{diagnostics.bayes_wilcox_test}
\alias{diagnostics.bayes_wilcox_test}
\title{Plots and prints diagnostics regarding the convergence of the model.}
\usage{
diagnostics.bayes_wilcox_test(fit)
}
\arguments{
\item{fit}{The output from a Bayesian First Aid model.}
}
\description{
Plots and prints diagnostics regarding the convergence of the model.
}
