% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test.R
\name{model.code.bayes_paired_wilcox_test}
\alias{model.code.bayes_paired_wilcox_test}
\title{Prints code that replicates the model you just ran.}
\usage{
model.code.bayes_paired_wilcox_test(fit)
}
\arguments{
\item{fit}{The output from a paired sample Bayesian Wilcox test model.}
}
\description{
This is good if you better want to understand how the model is
implemented or if you want to run a modified version of the code.
}
